oldcdk/StaticSite/requirements.txt
aws-cdk-lib==2.80.0
constructs>=10.0.0,<11.0.0
# aws-cdk.aws-s3>=1.202.0